[controller] Working user controller

+++ src/main/java/com/student/service/web/repository/AnnouncementRepository.java
@@ -5,5 +5,5 @@ import com.student.service.web.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AnnouncementRepository extends JpaRepository<Announcement, Long> {
+public interface AnnouncementRepository extends JpaRepository<Announcement, Integer> {
 }
